Try again to stabilize CI moving stateful suites out of async mode
