+++ hedera-node/src/main/resources/bootstrap.properties
@@ -47,7 +47,7 @@ balances.nodeBalanceWarningThreshold=0
 cache.records.ttl=180
 contracts.defaultLifetime=7890000
 contracts.localCall.estRetBytes=32
-contracts.maxGas=5000000
+contracts.maxGas=300000
 contracts.chainId=293
 contracts.throttle.throttleByGas=false
 contracts.maxRefundPercentOfGasLimit=20
