+++ test-clients/src/main/java/com/hedera/services/bdd/suites/contract/ContractCallLocalSuite.java
@@ -23,14 +23,12 @@ package com.hedera.services.bdd.suites.contract;
 import com.hedera.services.bdd.spec.HapiApiSpec;
 import com.hedera.services.bdd.spec.HapiSpecSetup;
 import com.hedera.services.bdd.spec.infrastructure.meta.ContractResources;
-import com.hedera.services.bdd.spec.utilops.UtilVerbs;
 import com.hedera.services.bdd.suites.HapiApiSuite;
 import com.hederahashgraph.api.proto.java.ResponseCodeEnum;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.math.BigInteger;
-import java.util.Arrays;
 import java.util.List;
 
 import static com.hedera.services.bdd.spec.HapiApiSpec.defaultHapiSpec;
@@ -49,7 +47,6 @@ import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.CONTRACT_DELET
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INSUFFICIENT_PAYER_BALANCE;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INSUFFICIENT_TX_FEE;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_CONTRACT_ID;
-import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.MAX_GAS_LIMIT_EXCEEDED;
 
 public class ContractCallLocalSuite extends HapiApiSuite {
 	private static final Logger log = LogManager.getLogger(ContractCallLocalSuite.class);
@@ -65,28 +62,14 @@ public class ContractCallLocalSuite extends HapiApiSuite {
 
 	@Override
 	protected List<HapiApiSpec> getSpecsInSuite() {
-		return allOf(
-				positiveSpecs(),
-				negativeSpecs()
-		);
-	}
-
-	private List<HapiApiSpec> negativeSpecs() {
-		return Arrays.asList(
-				deletedContract(),
-				invalidContractID(),
-				impureCallFails(),
-				insufficientFeeFails(),
-				lowBalanceFails(),
-				gasLimitOverMaxGasLimitFailsPrecheck()
-		);
-	}
-
-	private List<HapiApiSpec> positiveSpecs() {
-		return Arrays.asList(
-				vanillaSuccess(),
-				minChargeIsTXGasUsed(),
-				maxRefundIsMaxGasRefundConfiguredWhenTXGasPriceIsSmaller()
+		return List.of(new HapiApiSpec[] {
+						deletedContract(),
+						invalidContractID(),
+						impureCallFails(),
+						insufficientFeeFails(),
+						lowBalanceFails(),
+						vanillaSuccess(),
+				}
 		);
 	}
 
@@ -102,7 +85,7 @@ public class ContractCallLocalSuite extends HapiApiSuite {
 						contractCallLocal("parentDelegate", ContractResources.GET_CHILD_RESULT_ABI)
 								.has(resultWith().resultThruAbi(
 										ContractResources.GET_CHILD_RESULT_ABI,
-										isLiteralResult(new Object[]{BigInteger.valueOf(7L)})))
+										isLiteralResult(new Object[] { BigInteger.valueOf(7L) })))
 				);
 	}
 
@@ -190,49 +173,6 @@ public class ContractCallLocalSuite extends HapiApiSuite {
 						getAccountBalance("payer").logged(),
 						sleepFor(1_000L),
 						getAccountBalance("payer").logged()
-				);
-	}
-
-	private HapiApiSpec maxRefundIsMaxGasRefundConfiguredWhenTXGasPriceIsSmaller() {
-		return defaultHapiSpec("MaxRefundIsMaxGasRefundConfiguredWhenTXGasPriceIsSmaller")
-				.given(
-						UtilVerbs.overriding("contracts.maxRefundPercentOfGasLimit", "5"),
-						fileCreate("parentDelegateBytecode").path(ContractResources.DELEGATING_CONTRACT_BYTECODE_PATH),
-						contractCreate("parentDelegate").bytecode("parentDelegateBytecode")
-				).when(
-						contractCall("parentDelegate", ContractResources.CREATE_CHILD_ABI)
-				).then(
-						contractCallLocal("parentDelegate", ContractResources.GET_CHILD_RESULT_ABI).gas(300_000L)
-								.has(resultWith().gasUsed(285_000L)),
-						UtilVerbs.resetAppPropertiesTo("src/main/resource/bootstrap.properties")
-				);
-	}
-
-	private HapiApiSpec minChargeIsTXGasUsed() {
-		return defaultHapiSpec("MinChargeIsTXGasUsed")
-				.given(
-						UtilVerbs.overriding("contracts.maxRefundPercentOfGasLimit", "100"),
-						fileCreate("parentDelegateBytecode").path(ContractResources.DELEGATING_CONTRACT_BYTECODE_PATH),
-						contractCreate("parentDelegate").bytecode("parentDelegateBytecode")
-				).when(
-						contractCall("parentDelegate", ContractResources.CREATE_CHILD_ABI)
-				).then(
-						contractCallLocal("parentDelegate", ContractResources.GET_CHILD_RESULT_ABI).gas(300_000L)
-								.has(resultWith().gasUsed(5451)),
-						UtilVerbs.resetAppPropertiesTo("src/main/resource/bootstrap.properties")
-				);
-	}
-
-	private HapiApiSpec gasLimitOverMaxGasLimitFailsPrecheck() {
-		return defaultHapiSpec("GasLimitOverMaxGasLimitFailsPrecheck")
-				.given(
-						fileCreate("parentDelegateBytecode").path(ContractResources.DELEGATING_CONTRACT_BYTECODE_PATH),
-						contractCreate("parentDelegate").bytecode("parentDelegateBytecode"),
-						UtilVerbs.overriding("contracts.maxGas", "100")
-				).when().then(
-						contractCallLocal("parentDelegate", ContractResources.GET_CHILD_RESULT_ABI).gas(101L)
-								.hasCostAnswerPrecheck(MAX_GAS_LIMIT_EXCEEDED),
-						UtilVerbs.resetAppPropertiesTo("src/main/resource/bootstrap.properties")
 				);
 	}
